requirements.txt
## Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

## PDF Processing
PyPDF2==3.0.1
pdf2docx==0.5.6
pdf2image==1.16.3
img2pdf==0.5.1
Pillow==10.2.0
pdfplumber==0.11.0
openpyxl==3.1.2

## System utilities
python-magic==0.4.27
psutil==5.9.8

## Logging
loguru==0.7.2
